+++ lib/flutter_mono.dart
@@ -1,4 +1,5 @@
 library flutter_mono;
 
 export 'src/flutter_mono.dart';
-export 'src/views/mono_web_view.dart';
+export 'src/models/success_model.dart';
+export 'src/views/mono_view.dart';